Experimento II/seguridad/requirements.txt
Flask==2.1.2
gunicorn==20.1.0
Werkzeug==2.3.8
apscheduler==3.9.1
psycopg2-binary==2.9.7
